+++ CHANGELOG.md
@@ -38,6 +38,7 @@ This release adds support to OneView Rest API version 500 for the hardware varia
    - Server Profile Template
    - Storage Pool
    - Storage System
+   - Storage Templates
    - Switches
    - Unmanaged Devices
    - Uplink Set